+++ packages/cli-client/CHANGELOG.md
@@ -3,6 +3,94 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.3.61"></a>
+## [1.3.61](https://github.com/wireapp/wire-web-packages/tree/master/packages/cli-client/compare/@wireapp/cli-client@1.3.60...@wireapp/cli-client@1.3.61) (2018-04-17)
+
+
+
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+<a name="1.3.60"></a>
+## [1.3.60](https://github.com/wireapp/wire-web-packages/tree/master/packages/cli-client/compare/@wireapp/cli-client@1.3.59...@wireapp/cli-client@1.3.60) (2018-04-17)
+
+
+
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+<a name="1.3.59"></a>
+## [1.3.59](https://github.com/wireapp/wire-web-packages/tree/master/packages/cli-client/compare/@wireapp/cli-client@1.3.58...@wireapp/cli-client@1.3.59) (2018-04-17)
+
+
+
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+<a name="1.3.58"></a>
+## [1.3.58](https://github.com/wireapp/wire-web-packages/tree/master/packages/cli-client/compare/@wireapp/cli-client@1.3.57...@wireapp/cli-client@1.3.58) (2018-04-16)
+
+
+
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+<a name="1.3.57"></a>
+## [1.3.57](https://github.com/wireapp/wire-web-packages/tree/master/packages/cli-client/compare/@wireapp/cli-client@1.3.56...@wireapp/cli-client@1.3.57) (2018-04-14)
+
+
+
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+<a name="1.3.56"></a>
+## [1.3.56](https://github.com/wireapp/wire-web-packages/tree/master/packages/cli-client/compare/@wireapp/cli-client@1.3.55...@wireapp/cli-client@1.3.56) (2018-04-13)
+
+
+
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+<a name="1.3.55"></a>
+## [1.3.55](https://github.com/wireapp/wire-web-packages/tree/master/packages/cli-client/compare/@wireapp/cli-client@1.3.54...@wireapp/cli-client@1.3.55) (2018-04-13)
+
+
+
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+<a name="1.3.54"></a>
+## [1.3.54](https://github.com/wireapp/wire-web-packages/tree/master/packages/cli-client/compare/@wireapp/cli-client@1.3.53...@wireapp/cli-client@1.3.54) (2018-04-13)
+
+
+
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+<a name="1.3.53"></a>
+## [1.3.53](https://github.com/wireapp/wire-web-packages/tree/master/packages/cli-client/compare/@wireapp/cli-client@1.3.52...@wireapp/cli-client@1.3.53) (2018-04-12)
+
+
+
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+<a name="1.3.52"></a>
+## [1.3.52](https://github.com/wireapp/wire-web-packages/tree/master/packages/cli-client/compare/@wireapp/cli-client@1.3.51...@wireapp/cli-client@1.3.52) (2018-04-12)
+
+
+
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+<a name="1.3.51"></a>
+## [1.3.51](https://github.com/wireapp/wire-web-packages/tree/master/packages/cli-client/compare/@wireapp/cli-client@1.3.50...@wireapp/cli-client@1.3.51) (2018-04-12)
+
+
+
+
+**Note:** Version bump only for package @wireapp/cli-client
+
 <a name="1.3.50"></a>
 ## [1.3.50](https://github.com/wireapp/wire-web-packages/tree/master/packages/cli-client/compare/@wireapp/cli-client@1.3.49...@wireapp/cli-client@1.3.50) (2018-04-12)